Jmol.___JmolVersion="14.4.1_2016.01.06"
bug fix: labels, echos, and measurements disappear when z-shaded




Former-commit-id: fe48c15a19a9f24216b0e48c2e1d4f3a177f9486

+++ Jmol/src/org/jmol/api/JmolRendererInterface.java
@@ -260,7 +260,7 @@ public interface JmolRendererInterface extends JmolGraphicsInterface {
                                         double privateKey,
                                         Map<String, Object> params);
 
-  public abstract void plotImagePixel(int argb, int x, int y, int z, int shade, int bgargb, int width, int  height, int[] zbuf, Object pixel, int transpLog);
+  public abstract void plotImagePixel(int argb, int x, int y, int z, byte shade, int bgargb, int width, int  height, int[] zbuf, Object pixel, int transpLog);
 
   public abstract void plotPixelClippedP3i(P3i a);
